+++ src/main/java/com/bank/kata/service/AccountService.java
@@ -4,13 +4,14 @@ import com.bank.kata.exception.OperationException;
 import com.bank.kata.model.Transaction;
 import com.bank.kata.presentation.AccountPreview;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface AccountService {
 
-    public AccountPreview deposit(Long amount) throws OperationException;
+    public AccountPreview deposit(BigDecimal amount) throws OperationException;
 
-    public AccountPreview withdrawal(Long amount) throws OperationException;
+    public AccountPreview withdrawal(BigDecimal amount) throws OperationException;
 
     public List<Transaction> getTransactions();
